Elastic-ViT-Retrieval/requirements.txt
torch
torchvision
timm
transformers
PyYAML
tqdm
Pillow
pycocotools
safetensors
